+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Fixed native memory leak setting the value of a primary key (#3993).
 * Activated Realm's annotation processor on connectedTest when the project is using kapt (#4008).
+* Fixed bug, preventing Sync client to renew the access token (#4038) (#4039).
 
 ### Object Server API Changes (In Beta)