+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Fixed native memory leak setting the value of a primary key (#3993).
 * Activated Realm's annotation processor on connectedTest when the project is using kapt (#4008).
+* Fixed bug, preventing Sync client to renew the access token (#4038) (#4039).
+
+### Object Server API Changes (In Beta)
+
+* Exceptions thrown in error handlers are ignored but logged (#3559).
 
 ### Internal